requirements.txt
flask
azure-storage-blob
Werkzeug
gunicorn
pandas